refactor: simplify accept, discharge, and reject functions by utilizing runStatusAction
